Pass in GITHUB_SHA to container at build time
If we pass in a GITHUB_SHA_ARG at build time, this will be available
to the code later as process.env.GITHUB_SHA, which means that we
can put the build version in our markup.

That way, we will know what version of the app is running at any
given time.